changes to entypoint.sh

+++ entrypoint.sh
@@ -49,16 +49,24 @@ monitor_ups() {
 }
 
 # Ensure WOL configuration files are in place
-CONFIGS_DIR="/opt/scripts/configs"
+CONFIGS_DIR="scripts/configs"
 
 if [ ! -f "$CONFIGS_DIR/wol_clients.conf" ]; then
     echo "wol_clients.conf not found! Exiting..."
     exit 1
 fi
 
+# Ensure WOL script files are in place
+SCRIPTS_DIR="scripts/configs"
+
+if [ ! -f "$SCRIPTS_DIR/wol.sh" ]; then
+    echo "wol.sh not found! Exiting..."
+    exit 1
+fi
+
 # Copy WOL configuration files from the mounted volume to their respective locations
 cp "$CONFIGS_DIR/wol_clients.conf" /opt/scripts/wol_clients.conf
-cp "$CONFIGS_DIR/wol.sh" /opt/scripts/wol.sh
+cp "$SCRIPTS_DIR/wol.sh" /opt/scripts/wol.sh
 
 # Generate a file with the output of nut-scanner -U
 nut-scanner -U > /etc/nut/nut-scanner-output.txt
